+++ src/game/extras.ts
@@ -1,11 +1,10 @@
 import * as THREE from 'three';
 
-import { getRandom } from '../utils/lba';
+import { getRandom, getHtmlColor } from '../utils/lba';
 import { SpriteType } from './data/spriteType';
 import { loadSprite } from '../iso/sprites';
 import { addExtraToScene, removeExtraFromScene } from './scenes';
 import { clone } from 'lodash';
-import { getHtmlColor } from '../scene';
 import { MAX_LIFE } from './state';
 // import { createBoundingBox } from '../utils/rendering';
